Move some standalone functions into a separate helpers lib
